apps/worker-exchange-rates/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "basedguide-worker-exchange-rates"
version = "0.1.0"
description = "Exchange rates worker for BasedGuide"
authors = [
    {name = "BasedGuide", email = "contact@basedguide.com"},
]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "shared-eu @ file://../../packages/python",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]
src = ["src", "tests"]

[tool.black]
line-length = 100
